backend/pyproject.toml
[tool.poetry]
name = "react+fastapi-backend"
version = "0.1.0"
description = ""
authors = [" guilherme-biscaino  < guilhermebs.bs2@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.1"
alembic = "^1.14.1"
sqlalchemy = "^2.0.38"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
